publisher/requirements.txt
Flask==1.1.2
gunicorn==19.6.0
emoji==0.6.0
python-telegram-bot==13.0
